@@ -1,7 +1,7 @@
 require './lib/backup_pages'
 
 class ExportsController < ApplicationController
-  before_filter :required_login
+  before_action :required_login
 
   def zip
     zipfile = BackupPages.zip(current_user)
