@@ -1,5 +1,5 @@
 class SettingsController < ApplicationController
-  before_filter :required_login
+  before_action :required_login
 
   def show
     respond_to do |format|
