@@ -1,4 +1,4 @@
-class CreateChatrooms < ActiveRecord::Migration
+class CreateChatrooms < ActiveRecord::Migration[4.2]
   def change
     create_table :chatrooms do |t|
       t.string :key, :unique => true
