@@ -1,4 +1,4 @@
-class CreateChatMessages < ActiveRecord::Migration
+class CreateChatMessages < ActiveRecord::Migration[4.2]
   def change
     create_table :chat_messages do |t|
       t.integer :user_id
