@@ -1,4 +1,4 @@
-class CreatePageDates < ActiveRecord::Migration
+class CreatePageDates < ActiveRecord::Migration[4.2]
   def change
     create_table :page_dates do |t|
       t.integer :page_id
