@@ -1,4 +1,4 @@
-class CreateDropboxUsers < ActiveRecord::Migration
+class CreateDropboxUsers < ActiveRecord::Migration[4.2]
   def change
     create_table :dropbox_users do |t|
       t.integer :user_id, :unique => true
