@@ -1,4 +1,4 @@
-class CreateUserMessages < ActiveRecord::Migration
+class CreateUserMessages < ActiveRecord::Migration[4.2]
   def change
     create_table :user_messages do |t|
       t.integer :user_id
