@@ -1,4 +1,4 @@
-class AddActivedAtToUsers < ActiveRecord::Migration
+class AddActivedAtToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :actived_at, :timestamp
     add_index :users, [:actived_at]
