@@ -1,4 +1,4 @@
-class AddEmailToUsers < ActiveRecord::Migration
+class AddEmailToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :email, :string
   end
