@@ -1,4 +1,4 @@
-class AddSenderUserIdToUserMessages < ActiveRecord::Migration
+class AddSenderUserIdToUserMessages < ActiveRecord::Migration[4.2]
   def change
     add_column :user_messages, :sender_user_id, :integer
     add_index :user_messages, [:user_id, :sender_user_id, :created_at], :name => 'idx_user_messages_6'
