@@ -1,4 +1,4 @@
-class CreateFeedbacks < ActiveRecord::Migration
+class CreateFeedbacks < ActiveRecord::Migration[4.2]
   def change
     create_table :feedbacks do |t|
       t.integer :user_id
