@@ -1,4 +1,4 @@
-class AddUserAgentToFeedbacks < ActiveRecord::Migration
+class AddUserAgentToFeedbacks < ActiveRecord::Migration[4.2]
   def change
     add_column :feedbacks, :user_agent, :text
   end
