@@ -1,4 +1,4 @@
-class AddActiveToUsers < ActiveRecord::Migration
+class AddActiveToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :active, :boolean, default: true
     add_index :users, :active
