@@ -1,4 +1,4 @@
 class PageDate < ActiveRecord::Base
   belongs_to :page
-  belongs_to :user
+  belongs_to :user, optional: true
 end
