@@ -1,8 +1,10 @@
-# 
+# frozen_string_literal: true
+
+#
 WRIPE_VERSION = 11
 
-# Load the rails application.
-require File.expand_path('../application', __FILE__)
+# Load the Rails application.
+require_relative 'application'
 
-# Initialize the rails application.
-Wripe::Application.initialize!
+# Initialize the Rails application.
+Rails.application.initialize!
